WIP for new connection/channel system. + Added IConnection and IChannel (placeholder names). These classes are currently extremly voletile. I'm essentailly taking a look at my existing Connection.cpp and refactoring it into a simpler API. + Added m_numMessages field to Packet. This will (in the future) allow me to send multiple messages in a single packet. + Few other experiments peppered around.
More work is to be done. Here a motivating use case:

Let's consider a simple application where the client has 3 squares, all coloured blue. Each of these squares is an instansiation of Square.cpp.

When the user clicks on a square, the square instance will send a message to the server requesting a new colour. Upon receiving this request, the server will do some work and pick a random colour. The server will then send this selected colour in a new message back to the client.

The original square must now be updated with this new colour.

There are some core concepts required to acheive this simple use case:
1. A object needs to be able to "listen" for certain incoming messages (maybe even on certain channels from certain individuals, but this will be explored later). In our example, Square will listen for colour change messages.
2. Following point (1), we need a system that can distribute messages to their intended recipient. It follows that messages will need a concept of "identity". That identity will likely include information on:
   a) the peer who sent the message
   b) the instance id that created the message
   c) the instance id that will receive the message
3. Following point (2), we may want to allow for messages to be multicasted to multiple recipients. For example: every 20 seconds, our server will send every client's Square a new color. In this case, a & b are easy to create, but c is trickier.... how do we specify a list of instance ids to multiple peers?

Idea::

NetworkObject
- On creation, assigned a uuid. We can mandate that network objects can only be created via a call to some kind of factory. This would inforce the replication of the network object.

Our Square object would then inherit from this object.